Linux can now build libslang.dll
map.exe and adikted.dll were able to be cross-compiled on Linux before this commit. Howver, libslang.dll could not be cross-compiled.

With these changes, Linux can now cross-compile libslang.dll.

To be able to compile a DLL of the slang library with MinGW32 in Linus/WSL, mkmake must be used rather than ./configure.

mkmake.c is first compiled for Linux, and then used by:
A modifed version of the script in mkfiles/build.sh which is used to create the Makefiles for MinGW32, and then run make.

The parameters are the same as when building on Windows: WIN32 MINGW32 DLL

The key difference is adding $(CROSS_COMPILE) i.e. i686-w64-mingw32- to TOOL_PREFIX with the use of sed.