new libslang_private.rc injected to libslang.dll
Add some basic Windows Information to the libslang.dll.

Created /libslang_patches/libslang_private.rc with description, version and license info for S-Lang pre2.3.3-24

Amends `slang-src-mingw32-makefile.patch` to allow slang's makefile to create a .res file and inject it in to the dll.

Add the SLANG_RC_FOLDER variable to mapslang/Makefile, allowing it to send a location that contains a `libslang_private.rc`  to slang's make.