Character codes are now correctly sent to new S-Lang
This patch:
- Enables UTF-8 mode in libslang
- Casts the char sent to SLsmg_write_char() as an unsigned char
- **Doing this fixes the drawing of lines with mapslang** (Sending the charcode for any character from code page 850 works for me).

An `unsigned char` matches `SLuchar_Type` from slang, the precise reason that this works is unknown to me right now. Looking through SLsmg_write_chars() in slsmg.c should lead to an answer.

I'm not sure what determines the codepage here, or if it is compile time, or run time.

These changes do not cause an issue if you use map.exe (with these changes) with the older libslang.dll from 0.95